Refactor dialogfragment and adapter class
